doc_it/requirements.txt
Django>=3.0,
djangorestframework